+++ src/plone/restapi/tests/test_services_users.py
@@ -88,6 +88,19 @@ class TestUsersEndpoint(unittest.TestCase):
         self.assertEqual('Professor of Linguistics', noam.get('description'))  # noqa
         self.assertEqual('Cambridge, MA', noam.get('location'))
 
+    def test_list_users_without_being_manager(self):
+        noam_api_session = RelativeSession(self.portal_url)
+        noam_api_session.headers.update({'Accept': 'application/json'})
+        noam_api_session.auth = ('noam', 'password')
+
+        response = noam_api_session.get('/@users')
+        self.assertEqual(response.status_code, 401)
+
+    def test_list_users_as_anonymous(self):
+
+        response = self.anon_api_session.get('/@users')
+        self.assertEqual(response.status_code, 401)
+
     def test_add_user(self):
         response = self.api_session.post(
             '/@users',
@@ -297,6 +310,18 @@ class TestUsersEndpoint(unittest.TestCase):
         self.assertEqual('Professor of Linguistics', response.json().get('description'))  # noqa
         self.assertEqual('Cambridge, MA', response.json().get('location'))
 
+    def test_get_user_as_anonymous(self):
+        response = self.anon_api_session.get('/@users/noam')
+        self.assertEqual(response.status_code, 401)
+
+    def test_get_other_user_info_when_logged_in(self):
+        noam_api_session = RelativeSession(self.portal_url)
+        noam_api_session.headers.update({'Accept': 'application/json'})
+        noam_api_session.auth = ('noam', 'password')
+
+        response = noam_api_session.get('/@users/otheruser')
+        self.assertEqual(response.status_code, 401)
+
     def test_get_search_user_with_filter(self):
         response = self.api_session.post(
             '/@users',
@@ -326,6 +351,39 @@ class TestUsersEndpoint(unittest.TestCase):
         self.assertEqual(response.status_code, 200)
         self.assertEqual(len(response.json()), 1)
         self.assertEqual('howard', response.json()[0].get('id'))
+
+    def test_get_search_user_with_filter_as_anonymous(self):
+        response = self.api_session.post(
+            '/@users',
+            json={
+                "username": "howard",
+                "email": "howard.zinn@example.com",
+                "password": "peopleshistory"
+            },
+        )
+        transaction.commit()
+        response = self.anon_api_session.get(
+            '/@users',
+            params={'query': 'howa'}
+        )
+        self.assertEqual(response.status_code, 401)
+
+    def test_get_search_user_with_filter_as_unauthorized_user(self):
+        response = self.api_session.post(
+            '/@users',
+            json={
+                "username": "howard",
+                "email": "howard.zinn@example.com",
+                "password": "peopleshistory"
+            },
+        )
+        transaction.commit()
+        noam_api_session = RelativeSession(self.portal_url)
+        noam_api_session.headers.update({'Accept': 'application/json'})
+        noam_api_session.auth = ('noam', 'password')
+
+        response = noam_api_session.get('/@users', params={'query': 'howa'})
+        self.assertEqual(response.status_code, 401)
 
     def test_get_non_existing_user(self):
         response = self.api_session.get('/@users/non-existing-user')